requirements.txt
ccxt==4.1.83
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.24.4
aiohttp==3.9.1
asyncio-throttle==1.0.2
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
ta-lib==0.4.28
matplotlib==3.8.2
plotly==5.17.0
requests==2.31.0
websockets==12.0
pyyaml==6.0.1